Refactor bandwidth parser into focused modules
